requirements.txt
torch
transformers
datasets
evaluate
dill
filelock
inflection
nltk
numpy
Pillow
pycocoevalcap
pycocotools
tqdm
xxhash
bytecode
